backend/requirements.txt
# TurkDamga API — minimal çalışan iskelet (genişletme: TurkDamga-Backend.md)
fastapi>=0.109.0,<0.116.0
uvicorn[standard]>=0.27.0,<0.32.0
pydantic-settings>=2.2.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# İsteğe bağlı: DATABASE_URL tanımlıysa /health içinde db durumu kontrol edilir
sqlalchemy[asyncio]>=2.0.25,<2.1.0
asyncpg>=0.29.0,<0.31.0
httpx>=0.26.0,<0.29.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
email-validator>=2.0.0,<3.0.0
bcrypt>=3.2.0,<4.0.0
slowapi>=0.1.9,<0.2.0
